Refactored variable types.

+++ src/resistor_logic.c
@@ -3,19 +3,19 @@
 #include "resistor_logic.h"
 #include <math.h>
 
-const int CHARS_NUMERIC = 3;
-const int CHARS_MULTIPLIER = 7;
-const int CHARS_TOLERANCE = 7;
-const int CHARS_TEMP_COEFF = 3;
-const int CHARS_CALCULATION = 12;
+const uint8_t CHARS_NUMERIC = 3;
+const uint8_t CHARS_MULTIPLIER = 7;
+const uint8_t CHARS_TOLERANCE = 7;
+const uint8_t CHARS_TEMP_COEFF = 3;
+const uint8_t CHARS_CALCULATION = 12;
 
 const char BLANK_CALCULATION[] = "           "; // "nnn x 10^nn"
 const char BLANK_TOLERANCE[] = "       ";
 const char BLANK_TEMP_COEFF[] = "   ";
 
-const int INDEX_NUMERIC = 0;
-const int INDEX_MULTIPLIER = 4;
-const int INDEX_TOLERANCE = 0;
+const uint8_t INDEX_NUMERIC = 0;
+const uint8_t INDEX_MULTIPLIER = 4;
+const uint8_t INDEX_TOLERANCE = 0;
 
 const int NUMERIC_BANDS_PER_RESISTOR[6] = {-1, -1, 2, 2, 3, 3};
 const int MULTIPLIER_INDEX_PER_RESISTOR[6] = {-1, -1, 2, 2, 3, 3};
@@ -70,9 +70,12 @@ bool is_temp_coeff_colour(BandColour colour) {
            colour == BandBlue || colour == BandPurple || colour == BandGray;
 }
 
-BandColour
-    alter_resistor_band(ResistorType rtype, int band, BandColour current_colour, int direction) {
-    int colour = current_colour;
+BandColour alter_resistor_band(
+    ResistorType rtype,
+    uint8_t band,
+    BandColour current_colour,
+    int8_t direction) {
+    int8_t colour = current_colour;
     bool accepted = false;
     while(!accepted) {
         colour += direction;
@@ -87,9 +90,9 @@ BandColour
 }
 
 int decode_resistance_number(ResistorType rtype, BandColour colours[]) {
-    int bands = NUMERIC_BANDS_PER_RESISTOR[rtype - 1];
+    uint8_t bands = NUMERIC_BANDS_PER_RESISTOR[rtype - 1];
     int value = 0;
-    for(int b = 0; b < bands; b++) {
+    for(uint_fast8_t b = 0; b < bands; b++) {
         int pwr = bands - b - 1;
         int delta = ((int)pow(10.0, pwr)) * colours[b];
         value += delta;
@@ -111,7 +114,7 @@ char* calculate_decimal_places(double value) {
     while(end > decimalPoint && *(end - 1) == '0') // Skip trailing zeros
         end--;
 
-    formatter[2] = (int)(end - decimalPoint - 1) + '0';
+    formatter[2] = (uint8_t)(end - decimalPoint - 1) + '0';
     return formatter;
 }
 
@@ -119,7 +122,7 @@ void update_resistance_number(ResistorType rtype, BandColour colours[], char str
     double value = decode_resistance_number(rtype, colours);
     value *= resistor_multiplier;
     char* formatter = calculate_decimal_places(value);
-    int length = snprintf(NULL, 0, formatter, value);
+    uint8_t length = snprintf(NULL, 0, formatter, value);
     char* str = malloc(length + 1);
     snprintf(str, length + 1, formatter, value);
 
@@ -206,7 +209,7 @@ char* decode_resistance_multiplier(ResistorType rtype, BandColour colour) {
 }
 
 void update_resistance_multiplier(ResistorType rtype, BandColour colours[], char string[]) {
-    int multiplier_index = MULTIPLIER_INDEX_PER_RESISTOR[rtype - 1];
+    uint8_t multiplier_index = MULTIPLIER_INDEX_PER_RESISTOR[rtype - 1];
     calculate_resistance_decimal(rtype, colours[multiplier_index]);
     char* unit = decode_resistance_multiplier(rtype, colours[multiplier_index]);
     char* target = string + INDEX_MULTIPLIER;
@@ -281,7 +284,7 @@ char* decode_resistance_temp_coeff(BandColour colour) {
 
 void update_resistance_temp_coeff(ResistorType rtype, BandColour colours[], char string[]) {
     strcpy(string, BLANK_TEMP_COEFF);
-    int temp_coeff_index = TEMP_COEFF_INDEX_PER_RESISTOR[rtype - 1];
+    uint8_t temp_coeff_index = TEMP_COEFF_INDEX_PER_RESISTOR[rtype - 1];
     char* unit = decode_resistance_temp_coeff(colours[temp_coeff_index]);
     char* target = string + INDEX_TOLERANCE;
     strncpy(target, unit, CHARS_TEMP_COEFF);
